Apply aggregation to covasim-population.py
Introduction
apply aggregation in the project twice.

Abstract
Found those parts can apply aggregation to simplify code to understand. This is only being used as a practice of apply aggregate in the project which is complex. 

How were the changes made?
I created three new classes in population.py: "make_people_info", "pop_type", by changing and recode those variables, it made the project easy to add more parameters and further change. 

Why were the changes made?
While covid is spread, we need to consider a variety of variables about covid, but if the program is hard to change it will make the model less accurate by time passing.